# satellite/overlay: add changes to selected node benchmarks